Update rtl_entropy.c
This adds long options and a configuration file.  The only functionality changed is the addition of a --quiet / -q option that silences messages from rtl_entropy.  Set to 1, it eliminates the error messages from rejected packets, so that when streaming the output isn't overwhelmed by the errors.

The configuration file is called rtl_entropy in the program, and resides in /etc/sysconfig/rtl_entropy.