Update rtl_entropy.c
Clean up issues in configuration parsing related to multiple passes, especially for config files and output files.

Remove all option processing from parse_args, and do it after options are parsed.  So the configuration file is parsed exactly once, and any other effects are nullified.  

And the output file is processed exactly once, and is the last one found, either in the configuration file or on the command line.