Update rtl_entropy.c
The configuration file reader was choking on short options with values, because the token reader is destructive, and I hadn't taken that into account.  Fixed that.  

And set the optind to 1 at the start of the parse_args routine, so it will always start at the beginning of the options and can be reused without problem.

I think this takes care of problems with the configuration.  At least, I haven't found any more problems.